Remove Orator ORM and use repository pattern + pypika query builder

+++ src/lib/process_manager.py
@@ -8,17 +8,19 @@ import sys
 from typing import Optional, TypedDict
 from PyQt6 import QtCore
 from lib.browser_launcher.detect import Browser
-from lib.database import Database
-from models.data.client import Client
-from models.data.http_flow import HttpFlow
-from models.data.settings import Settings
-from models.data.websocket_message import WebsocketMessage
+from models.client import Client
+from models.http_flow import HttpFlow
+from models.websocket_message import WebsocketMessage
 from lib.proxy_handler import ProxyHandler
 from lib.paths import get_app_path
 from lib.utils import is_dev_mode
 from lib.browser_launcher.launch import launch_chrome_or_chromium, launch_firefox
 from lib.browser_launcher.browser_proc import BrowserProc
-from proxy.common_types import SettingsJson
+from models.http_response import HttpResponse
+from mitmproxy.common_types import ProxyRequest, ProxyResponse, ProxyWebsocketMessage, SettingsJson
+from repos.client_repo import ClientRepo
+from repos.http_flow_repo import HttpFlowRepo
+from repos.settings_repo import SettingsRepo
 
 class RunningProcess(TypedDict):
     client: Client
@@ -27,16 +29,17 @@ class RunningProcess(TypedDict):
     worker: Optional[BrowserProc]
 
 class ProcessManager(QtCore.QObject):
-    flow_created = QtCore.pyqtSignal(HttpFlow)
-    flow_updated = QtCore.pyqtSignal(HttpFlow)
+    proxy_request = QtCore.pyqtSignal(HttpFlow)
+    proxy_response = QtCore.pyqtSignal(HttpFlow)
+    proxy_ws_message = QtCore.pyqtSignal(WebsocketMessage)
     flow_intercepted = QtCore.pyqtSignal(HttpFlow)
-    websocket_message_created = QtCore.pyqtSignal(WebsocketMessage)
     proxy_started = QtCore.pyqtSignal(int)
 
     clients_changed = QtCore.pyqtSignal()
     recording_changed = QtCore.pyqtSignal(bool)
     intercept_changed = QtCore.pyqtSignal(bool)
 
+    app_path: str
     proxy_handler: ProxyHandler
     processes: list[RunningProcess]
     recording_enabled: bool
@@ -52,9 +55,9 @@ class ProcessManager(QtCore.QObject):
             raise Exception("Calling ProcessManager.get_instance() when there is not instance!")
         return ProcessManager.__instance
 
-    def __init__(self, src_path):
+    def __init__(self, app_path: str):
         super().__init__()
-        self.init(src_path)
+        self.init(app_path)
 
         # Virtually private constructor.
         if ProcessManager.__instance is not None:
@@ -63,20 +66,18 @@ class ProcessManager(QtCore.QObject):
             ProcessManager.__instance = self
     # /Singleton method stuff
 
-    def init(self, src_path):
-        self.src_path = src_path
+    def init(self, app_path: str):
+        self.app_path = app_path
         self.processes = []
         self.threadpool = QtCore.QThreadPool()
 
         self.proxy_handler = ProxyHandler(self)
         self.proxy_handler.start()
-        self.set_settings(Settings.get().parsed())
+        self.set_settings(SettingsRepo().get_settings().json)
 
-        self.proxy_handler.signals.flow_created.connect(self.flow_created)
-        self.proxy_handler.signals.flow_updated.connect(self.flow_updated)
-        self.proxy_handler.signals.flow_intercepted.connect(self.flow_intercepted)
-        self.proxy_handler.signals.websocket_message_created.connect(self.websocket_message_created)
-        self.proxy_handler.signals.proxy_started.connect(self.proxy_started)
+        self.proxy_handler.signals.proxy_request.connect(self.proxy_request_slot)
+        self.proxy_handler.signals.proxy_response.connect(self.proxy_response_slot)
+        self.proxy_handler.signals.proxy_ws_message.connect(self.proxy_ws_message_slot)
         self.proxy_handler.signals.proxy_started.connect(self.proxy_was_launched)
 
         self.recording_enabled = True
@@ -100,6 +101,8 @@ class ProcessManager(QtCore.QObject):
 
     def close_proxy(self, client: Client):
         process = [p for p in self.processes if p['client'].id == client.id and p['type'] == 'proxy'][0]
+        if process['process'] is None:
+            return
         pid = process['process'].pid
         print(f'[ProcessManager] killing process {pid}')
         os.kill(pid, signal.SIGTERM)
@@ -108,6 +111,8 @@ class ProcessManager(QtCore.QObject):
 
     def close_browser(self, client: Client):
         process = [p for p in self.processes if p['client'].id == client.id and p['type'] == 'browser'][0]
+        if process['worker'] is None:
+            return
         process['worker'].kill()
         self.processes.remove(process)
 
@@ -122,8 +127,8 @@ class ProcessManager(QtCore.QObject):
 
         self.processes.remove(browser_processes[0])
 
-        database = Database.get_instance()
-        database.db.table('clients').where('id', client.id).update(open=False)
+        client.open = False
+        ClientRepo().save(client)
         self.clients_changed.emit()
 
     def launch_client(self, client: Client, client_info: Browser, settings: SettingsJson):
@@ -135,13 +140,12 @@ class ProcessManager(QtCore.QObject):
             self.launch_browser(client, browser_command)
 
     def proxy_was_launched(self, client_id: int):
-        client = Client.find(client_id)
+        client = ClientRepo().find(client_id)
         if client is None:
             return
 
-        database = Database.get_instance()
-        database.db.table('clients').where('id', client.id).update(open=True)
-
+        client.open = True
+        ClientRepo().save(client)
         self.clients_changed.emit()
 
     def close_client(self, client: Client):
@@ -151,7 +155,7 @@ class ProcessManager(QtCore.QObject):
             self.close_proxy(client)
 
         client.open = False
-        client.save()
+        ClientRepo().save(client)
 
     def launch_browser(self, client: Client, browser_command):
         if client.type in ['chrome', 'chromium']:
@@ -166,17 +170,17 @@ class ProcessManager(QtCore.QObject):
         self.processes.append({'client': client, 'type': 'browser', 'worker': worker, 'process': None})
 
     def launch_proxy(self, client: Client, settings: SettingsJson):
-        app_path = str(get_app_path())
-        print(f"[ProcessManager] Launching proxy, app_path: {app_path}")
+        print(f"[ProcessManager] Launching proxy, app_path: {self.app_path}")
 
         recording_enabled = 1 if self.recording_enabled else 0
         intercept_enabled = 1 if self.intercept_enabled else 0
         settings_json_b64 = base64.b64encode(bytes(json.dumps(settings), 'utf-8')).decode('utf-8')
+        args_str = f'{client.id} {recording_enabled} {intercept_enabled} {settings_json_b64}'
 
         if is_dev_mode():
-            proxy_command = f'{sys.executable} {app_path}/proxy {client.proxy_port} {client.id} _ {recording_enabled} {intercept_enabled} {settings_json_b64}'
+            proxy_command = f'mitmdump -s {self.app_path}/src/mitmproxy/addon.py -p {client.proxy_port} --set confdir=./include {args_str}'
         else:
-            proxy_command = f'{app_path}/pntest_proxy {client.proxy_port} {client.id} {app_path}/include {recording_enabled} {intercept_enabled} {settings_json_b64}'
+            proxy_command = f'{self.app_path}/pntest/mitmdump -s {self.app_path}/pntest/addon.py -p {client.proxy_port} --set confdir=./include {args_str}'
 
         print(proxy_command)
         current_env = os.environ.copy()
@@ -187,7 +191,7 @@ class ProcessManager(QtCore.QObject):
         )
         self.processes.append({'client': client, 'type': 'proxy', 'process': process, 'worker': None})
 
-    def forward_flow(self, flow, intercept_response):
+    def forward_flow(self, flow: HttpFlow, intercept_response: bool):
         self.proxy_handler.forward_flow(flow, intercept_response)
 
     def forward_all(self):
@@ -216,3 +220,37 @@ class ProcessManager(QtCore.QObject):
 
     def set_settings(self, settings: SettingsJson) -> None:
         self.proxy_handler.set_settings(settings)
+
+    def proxy_request_slot(self, proxy_request: ProxyRequest):
+        flow = HttpFlow.from_proxy_request(proxy_request)
+        HttpFlowRepo().save(flow)
+
+        self.proxy_request.emit(flow)
+        if proxy_request['intercepted']:
+            self.flow_intercepted.emit(flow)
+
+    def proxy_response_slot(self, proxy_response: ProxyResponse):
+        flow = HttpFlowRepo().find_by_uuid(proxy_response['flow_uuid'])
+        if flow is None:
+            return
+        response = HttpResponse.from_state(proxy_response)
+        flow.add_response(response)
+        HttpFlowRepo().save(flow)
+
+        self.proxy_response.emit(flow)
+        if proxy_response['intercepted']:
+            self.flow_intercepted.emit(flow)
+
+    def proxy_ws_message_slot(self, proxy_ws_message: ProxyWebsocketMessage):
+        flow = HttpFlowRepo().find_by_uuid(proxy_ws_message['flow_uuid'])
+        if flow is None:
+            return
+
+        ws_message = WebsocketMessage.from_state(proxy_ws_message)
+        flow.add_ws_message(ws_message)
+        HttpFlowRepo().save(flow)
+
+        self.proxy_ws_message.emit(ws_message)
+        if proxy_ws_message['intercepted']:
+            flow.intercept_websocket_message = True
+            self.flow_intercepted.emit(flow)
